package/libsepol: fix help text for BR2_PACKAGE_LIBSEPOL_POLICY_VERSION
The help text was wrong, as it didn't match the actual default values
we were specifying. Indeed, when we specify:

        default 31 if BR2_TOOLCHAIN_HEADERS_AT_LEAST_4_13
        default 30 if BR2_TOOLCHAIN_HEADERS_AT_LEAST_4_3

It means that the policy version 30 is supported starting from Linux
4.3 included, and that 31 is supported from Linux 4.13 included.

So we shouldn't have:

	  > 4.3 <= 4.13    30
	  > 4.13           31

but:

	  >= 4.3 < 4.13    30
	  >= 4.13          31

This patch fixes that for all versions.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
(cherry picked from commit 67d7705a9aa58886e8a8777fe4c95746c1fb5c96)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>