package/python-networkx: install for host-python3
This Python module only works for Python 3: the target variant has a
"depends on BR2_PACKAGE_PYTHON3".

However today in Buildroot, when no target python is selected, or when
BR2_PACKAGE_PYTHON=y, all host python modules are installed for
host-python, i.e Python 2. But this module won't install in Python 2,
so let's force its host variant to be installed with Python 3 on the
host. Of course, for that to work, its dependency must also be built
for host-python3, so we change it to the newly introduced
host-python3-decorator package.

Fixes:

  http://autobuild.buildroot.net/results/46a89bd6cd0d0b896b28010db287068309e7a43c/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>