defconfigs: use BR2_PACKAGE_HOST_UBOOT_TOOLS_BOOT_SCRIPT_
For defconfigs that create a boot script, add the host package
uboot-tools and update the script variable names.

Add BR2_PACKAGE_HOST_UBOOT_TOOLS=y just before the existing
BR2_TARGET_UBOOT_BOOT_SCRIPT option. Note that for the defconfigs that
are not manually formatted (i.e. just the result of 'make
savedefconfig'), this puts them in the wrong place. However, it's easier
this way, and it's more consistent in general to have the definition of
the script close to BR2_TARGET_UBOOT itself.

Special cases:
- nanopi_* had the definition of the _SCRIPT variable in a weird place.
  These are moved to close to BR2_TARGET_UBOOT.
- Same for orangepi_plus.
- orangpi_pc_* already had BR2_PACKAGE_HOST_UBOOT_TOOLS=y so it's not
  added there.

Related to db2b6841263c3062fa1661dcfda3110415779fa2

Cc: Arnout Vandecappelle <arnout@mind.be>
Signed-off-by: Matthew Weber <matthew.weber@rockwellcollins.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>